refactor: move statistic logic to separate file
